# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de base en laravel",
        "description": "GabomaPay",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://gabomapay.com/api"
        }
    ],
    "paths": {
...
        "/admin/posts/store": {
            "post": {
                "tags": [
                    "ADMINISTRATION__PUBLICATION"
                ],
                "summary": "Ajouter une publication",
                "operationId": "65cf97fc6835798bd59ee2ceb01c7b80",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "Un exemple de description"
                                    },
                                    "image": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/posts/show/{id}": {
            "get": {
                "tags": [
                    "ADMINISTRATION__PUBLICATION"
                ],
                "summary": "Renvoie les informations d'une publications par son ID",
                "operationId": "b71295fbfc89400f547c1295507ae6bc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 18
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/posts/update": {
            "post": {
                "tags": [
                    "ADMINISTRATION__PUBLICATION"
                ],
                "summary": "Modifie une publication",
                "operationId": "7c3adf142329901e99a0c1e59806a9da",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Un exemple de description"
                                    },
                                    "image": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/posts/activate/{id}": {
            "get": {
                "tags": [
                    "ADMINISTRATION__PUBLICATION"
                ],
                "summary": "Active ou désactive l'état d'une publication par son ID",
                "operationId": "ec15a8a1c9758db9c390fd4c4652fee0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 18
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/admin/taux/getaux/{id}": {
            "get": {
                "tags": [
                    "ADMINISTRATION__TAUX"
                ],
                "summary": "Récupére un taux d'échange",
                "operationId": "06e8f9fd01cfe5203562f645ad75cfee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 18
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/taux/update": {
            "post": {
                "tags": [
                    "ADMINISTRATION__TAUX"
                ],
                "summary": "Modifie le taux d'échange",
                "operationId": "12a907a72689599e0906834599b3cd47",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "taux_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "taux_amount": {
                                        "type": "float",
                                        "example": 572.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/admin/users/show/{id}": {
            "get": {
                "tags": [
                    "ADMINISTRATION__UTILISATEUR"
                ],
                "summary": "Renvoie les informations d'un utilisateur par son ID",
                "operationId": "085e470d12a568ddd14f5a29270456ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 18
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/users/activate/{id}": {
            "get": {
                "tags": [
                    "ADMINISTRATION__UTILISATEUR"
                ],
                "summary": "Active ou désactive l'état d'un utilisateur par son ID",
                "operationId": "f43481b9e665f9e8f752b7c8b8dd8c06",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 18
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "bearerFormat": "JWT"
            },
            "0": []
        }
    }
}